requirements.txt
absl-py==1.0.0
appnope==0.1.2
astunparse==1.6.3
backcall==0.2.0
cachetools==5.1.0
certifi==2020.12.5
charset-normalizer==2.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.23
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.11.0
torch-tb-profiler==0.4.0
torchvision==0.12.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
